Refactor BuyProductListView to utilize BuyProductListService for filtering and pagination, improving code clarity and maintainability. Removed the ParseSortType method and replaced it with SortTypeConverterService for sorting functionality.
